pyaiot/gateway/mqtt: reorganize files and object names
